requirements.txt
uvicorn~=0.28.0
fastapi~=0.110.0
starlette~=0.36.3
pymongo~=4.6.2
python-dotenv~=1.0.1
redis~=5.0.3
pydantic~=2.6.4
bcrypt~=4.0.1
email-validator~=2.1.1
python-multipart~=0.0.9
passlib~=1.7.4
python-jose~=3.3.0
PyJWT~=2.8.0
motor~=3.3.2
aiohttp~=3.9.3
httpx~=0.27.0
requests~=2.31.0
aiosmtplib~=3.0.1
pydantic[email]
azure-storage-blob
azure-communication-email
azure.core